+++ src/test/kotlin/org/jetbrains/intellij/platform/gradle/tasks/PublishPluginTaskTest.kt
@@ -204,6 +204,7 @@ class PublishPluginTaskTest : IntelliJPluginTestBase() {
             buildFile write //language=kotlin
                     """
                     intellijPlatform {
+                        pluginConfiguration.ideaVersion.sinceBuild = "$nativeVariantsSinceBuild"
                         nativeVariants {
                             enabled = true
                         }